pyproject.toml
[tool.poetry]
name = "pytest-kind"
version = "0.1.0"
description = "Kubernetes test support with KIND for pytest"
authors = ["Henning Jacobs <henning@jacobs1.de>"]
license = "GPL-3.0+"
readme = "README.md"
repository = "https://codeberg.org/hjacobs/pytest-kind"
classifiers = [
    "Framework :: Pytest"
]

[tool.poetry.plugins.pytest11]
pytest-kind = "pytest_kind.plugin"

[tool.poetry.dependencies]
python = "^3.7"
pykube-ng = "^0.30.0"

[tool.poetry.dev-dependencies]
pytest = "^5.1"
pytest-cov = "^2.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
